+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.2-alpha.9",
-  "contentSha256_12": "e2289351b2d7",
+  "pkgVersion": "1.0.2-alpha.10",
+  "contentSha256_12": "9eb243da5966",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.2-alpha.9";
-export const contentSha256_12 = "e2289351b2d7";
+export const pkgVersion = "1.0.2-alpha.10";
+export const contentSha256_12 = "9eb243da5966";
 export const file = "nodalia-cards.bundle.js";